Updated user view and header, added footer

+++ pages/author/author_dashboard.php
@@ -17,7 +17,25 @@ require_once HTML_HEADER;
         height: 100px;
     }
 
+
     /* table */
+    .author-novels-header {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        padding: 20px 0;
+
+        button {
+            display: flex;
+            gap: 10px;
+            padding: 10px;
+            border-radius: 5px;
+            border: none;
+            font-weight: bold;
+            border: 1px solid black;
+        }
+    }
+
     tr {
         grid-template-columns: auto 100px 200px 100px 200px;
     }
@@ -27,40 +45,22 @@ require_once HTML_HEADER;
         justify-content: flex-start;
         align-items: center;
         gap: 20px;
-        height: 100px;
+        height: 150px;
     }
 
     .novel-operations {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        gap: 10px;
-
-        button {
-            justify-content: center;
-            width: 100%;
-            padding: 5px;
-        }
-
-        .delete-novel-button {
-            background-color: var(--danger-color);
-        }
-
-        .delete-novel-button:hover {
-            background-color: var(--danger-color-hover);
-        }
+        gap: 20px;
     }
 </style>
 <script>
     async function loadNovels() {
-        const tableBody = document.querySelector('tbody');
-        tableBody.innerHTML = '';
         const authorNovels = await axios.get('<?php echo NOVEL_API; ?>', {
             params: {
-                nv_author_id: <?php echo $_SESSION['author_id']; ?>
+                // nv_author_id: <?php echo $_SESSION['author_id']; ?>
+                nv_author_id: 6
             }
         });
+        const tableBody = document.querySelector('tbody');
         for (const novel of authorNovels.data) {
             tableBody.innerHTML += `
                 <tr id="${novel.nv_novel_id}">
@@ -78,26 +78,22 @@ require_once HTML_HEADER;
                 <td class="novel-publish-date">${novel.nv_publish_date}</td>
                 <td class="novel-rating">idk</td>
                 <td class="novel-operations">
-                    <button onclick="window.location.href='<?php echo AUTHOR_NOVEL_VIEW_PAGE; ?>?novel_id=${novel.nv_novel_id}'" class="view-novel-button">
-                        View Novel
-                    </button>
-                    <button data-novel-id="${novel.nv_novel_id}" data-novel-title="${novel.nv_novel_title}" class="delete-novel-button">
-                        Delete Novel
-                    </button>
+                    <a href="<?php echo AUTHOR_EDIT_NOVEL_PAGE; ?>?novel_id=${novel.nv_novel_id}">
+                        <i class="fa-solid fa-eye"></i>
+                    </a>
+                    <a class="delete-novel" data-novel-id="${novel.nv_novel_id}" data-novel-title="${novel.nv_novel_title}">
+                        <i class="fa-solid fa-trash"></i>
+                        </a>
                     </td>
                 </tr>
             `;
         }
     }
 
-    // <a href="<?php echo AUTHOR_NOVEL_VIEW_PAGE; ?>?novel_id=${novel.nv_novel_id}">
-    //                     <i class="fa-solid fa-eye"></i>
-    //                 </a>
-    //                 <a class="delete-novel" data-novel-id="${novel.nv_novel_id}" data-novel-title="${novel.nv_novel_title}">
-    //                     <i class="fa-solid fa-trash"></i>
-    //                     </a>
+    async function deleteNovel(novelId, novelName) {
+        const deleteBook = await confirm(`Are you sure you want to delete the following novel?\n\"${novelName}\"`);
+        if (!deleteBook) return;
 
-    async function deleteNovel(novelId) {
         const novelDeletionResponse = await axios.delete('<?php echo NOVEL_API; ?>', {
             params: {
                 nv_novel_id: novelId
@@ -106,32 +102,20 @@ require_once HTML_HEADER;
 
         if (novelDeletionResponse.status != 200) {
             alert('Failed to delete novel');
-            return false;
+            return;
         }
 
-        return true;
+        window.location.reload();
     }
 
     function enableDeleteNovelButtons() {
-        const deleteButtons = document.querySelectorAll('.delete-novel-button');
+        const deleteButtons = document.querySelectorAll('.delete-novel');
         deleteButtons.forEach(button => {
-            button.addEventListener('click', async (e) => {
+            button.addEventListener('click', (e) => {
                 e.preventDefault();
                 const novelId = e.currentTarget.dataset.novelId;
                 const novelTitle = e.currentTarget.dataset.novelTitle;
-                const userConfirmation = confirm(`Are you sure you want to delete the following novel?\n"${novelTitle}"\nThis action cannot be undone.`);
-                if (!userConfirmation) {
-                    return;
-                }
-
-                const isNovelDeleted = await deleteNovel(novelId);
-                if (!isNovelDeleted) {
-                    alert('Failed to delete novel');
-                    return;
-                }
-
-                alert('Novel deleted successfully');
-                await loadNovels();
+                deleteNovel(novelId, novelTitle);
             });
         });
     }
@@ -162,9 +146,9 @@ require_once HTML_HEADER;
 
         </div>
         <div class="author-novels">
-            <div class="table-title-row">
+            <div class="author-novels-header">
                 <h2>Published Novels</h2>
-                <button id="create-novel-button" on><i class="fa-solid fa-plus"></i>Create New Novel</button>
+                <button id="create-novel-button"><i class="fa-solid fa-plus"></i>Create New Novel</button>
             </div>
             <table>
                 <thead>
